src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/shadow-element-distributed-nodes-expected.txt

CONSOLE ERROR: line 87: Uncaught TypeError: element.createShadowRoot is not a function
Tests for a shadow element's getDistributedNodes().

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


getDistributedNodes() for "shadow" should return []
PASS distributedNodes.length is expectedDistributedNodes.length
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
